(junit) Split isAssignmentRequired tests into separate class
